darajati: Made is possible for coordinator to see what instructor sees
This will effect the whole system as in who ever is coordinator can view/edit
attendance, enrollment, grades

also removed the user profile where it fixed a performance issue.
The issue is using the def __str__(self)
and returning self.user would make 1 query per user which made the loading so
huge up to 7000 query for the attendance page and even in the admin page.