a huge code revamp, reusable code, better naming style
"I should not do such huge changes in a signel revision".
The point of this i was maintaining a code style which was somehow taken
from the django docs and best practices but i stop doing it in the middle
of the development now i have got back to the same style all the code should
be now documented and reduced. This made it easy to fix a lot of performance
issues.