double or n-number of marking can happen for a grade fragment
This is to test the new app called exam - we might need a better name.
The purpose of this app is there is a way to randomize students to take 2 or
exams related to one of the grade fragments.

This app should allow the creation of rooms and create exams for these rooms
also assign multiple instructor to these exams. In the StudentGrade model
we will assign the marker field to one of these assignments in the Assignment
model. where only this instructor is allowed to enter this mark or a
coorinator. This will be revised many times since we have 2 solutions for now.