removal_of_old_condition: only sections that has not been created will be sync
this means sections that have not been created will be sync with their periods
which is very dependent on the web-service and the non-created-section.

this should solve if there is a "data issue in the registrat data" and it was
fixed while the sections has been created it should create the periods.

|banner-integration|